define the hazmat module and move the new guts APIs there
